+++ forms/src/index.ts
@@ -17,4 +17,5 @@ export * from './modules/numberInput.module';
 export * from './modules/hasError.module';
 export * from './modules/requiredClass.module';
 export * from './modules/formPipes';
+export * from './modules/errorMessage';
 export * from './misc/formModel';