+++ forms/src/modules/errorMessage/components/index.ts
@@ -1,0 +1,1 @@
+export * from './groupErrors/groupErrors.component';