+++ forms/src/modules/errorMessage/directives/index.ts
@@ -1,0 +1,3 @@
+export * from './errorMessage/errorMessage.directive';
+export * from './groupErrorsTemplate/groupErrorsTemplate.directive';
+export * from './groupErrorsTemplate/groupErrorTemplate.context';