+++ forms/src/modules/errorMessage/pipes/index.ts
@@ -1,0 +1,1 @@
+export * from './withErrorMessage/withErrorMessage.pipe';