+++ src/pipes/index.ts
@@ -3,3 +3,4 @@ export * from './isPresent/isPresent.pipe';
 export * from './isNaN/isNaN.pipe';
 export * from './localize/localize.pipe';
 export * from './mergeCssClasses/mergeCssClasses.pipe';
+export * from './keys/keys.pipe';